requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
diffractsim @ git+https://github.com/aliceduque/diffractsim.git@002ddf37251551b4b8b2844af8e6a209fb6ceca6
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
idna==3.11
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==26.0
pillow==10.4.0
progressbar==2.5
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
soupsieve==2.7
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
